Add styling to settings page
