+++ src/App.js
@@ -8,8 +8,8 @@ import GlobalStyle from './GlobalStyle'
 
 export default function App() {
   const [name, setName] = useState('there')
-  const [interval, setInterval] = useState('5')
-  const [breakInterval, setBreakInterval] = useState('10')
+  const [interval, setInterval] = useState('25')
+  const [breakInterval, setBreakInterval] = useState('5')
   const [currentTodo, setCurrentTodo] = useState('')
   const [todos, setTodos] = useState([])
